% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_habitatTool.R
\name{cleanHabitatNames}
\alias{cleanHabitatNames}
\title{Habitat name cleaning function}
\usage{
cleanHabitatNames(name)
}
\arguments{
\item{name}{a character vector. A name from the 'habtype' column in the 'habitatPools' dataframe exported in this package, to be mapped to the correct utf-8 escaped danish spelling.}
}
\value{
a character vector of the same length as 'name'.
}
\description{
A function that takes one or more habitat names from the included habitatPools dataframe and inserts the correct local danish characters.
}
