% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DFV_pensum_w_ID}
\alias{DFV_pensum_w_ID}
\title{DFV_pensum_w_ID}
\format{
## `DFV_pensum_w_ID`
}
\usage{
data(DFV_pensum_w_ID)
}
\description{
Internal app dataframe. Exposed for transparency.
}
\keyword{datasets}
