% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{habitatPools}
\alias{habitatPools}
\title{habitatPools}
\format{
## `habitatPools`
}
\usage{
data(habitatPools)
}
\description{
Internal app dataframe. Exposed for transparency.
}
\keyword{datasets}
