% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{observations}
\alias{observations}
\title{observations}
\format{
## `observations`
}
\usage{
data(observations)
}
\description{
Internal app dataframe. Exposed for transparency.
}
\keyword{datasets}
