% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_habitatTool.R
\name{replace_eoa}
\alias{replace_eoa}
\title{replace_eoa}
\usage{
replace_eoa(str)
}
\arguments{
\item{str, }{a character vector}
}
\value{
a character vector with danish letters replaced by eoa
}
\description{
replace_eoa
}
