% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{speciesMeta}
\alias{speciesMeta}
\title{speciesMeta}
\format{
## `speciesMeta`
}
\usage{
data(speciesMeta)
}
\description{
Internal app dataframe. Exposed for transparency.
}
\keyword{datasets}
