% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{observations}
\alias{observations}
\alias{arterDKMeta}
\title{arterDKMeta}
\format{
## `arterDKMeta`
}
\usage{
data(arterDKMeta)
}
\description{
Internal app dataframe. Exposed for transparency.
}
\keyword{datasets}
